web/requirements.txt
flask
pandas
scikit-learn
joblib
googlemaps
python-dotenv